chord: Page Modals 폴더 Components로 이동
