+++ src/Components/Modals/index.ts
@@ -1,1 +1,5 @@
+export { default as StoreOpenCloseModal } from "./StoreOpenCloseModal";
+export { default as StoreDeleteModal } from "./StoreDeleteModal";
 export { default as Modal } from "./Modal";
+export { ProductDeleteModal } from "./ProductDeleteModal";
+export { ProductUpdateModal, ProductUpdateForm } from "./ProductUpdateModal";
