+++ src/Components/index.ts
@@ -3,7 +3,7 @@ export { Chart } from "./Chart";
 export { InputDefault, LabelDefault, TextareaDefault } from "./Form";
 export { Images, Noimage } from "./Images";
 export { Loading, LoadingBall } from "./Loading";
-export { Modal } from "./Modals";
+
 export { Nav } from "./Nav";
 export {
   PageHeader,
@@ -13,3 +13,12 @@ export {
   LandingLayout,
   OrderLayout
 } from "./Layouts";
+
+export {
+  StoreOpenCloseModal,
+  StoreDeleteModal,
+  ProductDeleteModal,
+  ProductUpdateModal,
+  ProductUpdateForm,
+  Modal
+} from "./Modals";
