--- src/Page/Admin/Modals/index.ts
@@ -1,2 +1,0 @@
-export { default as StoreOpenCloseModal } from "./StoreOpenCloseModal";
-export { default as StoreDeleteModal } from "./StoreDeleteModal";
